refactor(processing): split decoding and blending out of the streaming assembler
streaming_assembler.py was 998 lines, two under the 1000-line hard limit, and
held three responsibilities: reading clips into frames, composing frames into
the output stream, and driving FFmpeg. Split along the pipe:

- streaming_frame_decoder.py: FrameDecoder and make_decoder() — the per-clip
  FFmpeg filter chain (rotation, privacy blur, fill mode, HDR transfer,
  captions) and the raw-frame/audio output. Moved verbatim; _make_decoder is
  now public because title_inserter already imported it across module lines.
- streaming_frame_blender.py: FrameBlender — writes frames to a FrameSink
  (Protocol) and crossfades across clip boundaries. It owns the blend buffer
  plus the frame count, progress reporting and preview throttling that used
  to be threaded through eleven-argument helper calls.
- streaming_assembler.py keeps StreamingEncoder and the two entry points, and
  composes the blender with the encoder injected as its sink.

Pure move: no behaviour change. _encode_clip_sequence loses five parameters
because the blender carries that state, and the two identical decoder call
sites collapse into one local closure.
