+++ src/immich_memories/processing/title_inserter.py
@@ -72,17 +72,17 @@ class TitleInserter:
             create_assembly_context,
         )
         from immich_memories.processing.clip_encoder import encoder_args_for_plan
-        from immich_memories.processing.streaming_assembler import _make_decoder
+        from immich_memories.processing.streaming_frame_decoder import make_decoder
 
         plan = self.settings.encoding_plan
         output_path = output_dir / f"first_clip_processed.{plan.container}"
         ctx = create_assembly_context(self.settings, self.prober, clips, target_w, target_h)
 
-        # WHY: _make_decoder applies the EXACT same filter chain as the
+        # WHY: make_decoder applies the EXACT same filter chain as the
         # streaming assembler: rotation, scale_mode (blur bg), HDR conversion,
         # resolution, fps, SAR. The output is pixel-identical to what the
         # assembler will produce for this clip.
-        decoder = _make_decoder(
+        decoder = make_decoder(
             clips[0],
             0,
             target_w,
@@ -304,14 +304,14 @@ class TitleInserter:
 
         from immich_memories.processing.assembly_engine import create_assembly_context
         from immich_memories.processing.clip_encoder import encoder_args_for_plan
-        from immich_memories.processing.streaming_assembler import _make_decoder
+        from immich_memories.processing.streaming_frame_decoder import make_decoder
 
         clip = clips[-1]
         plan = self.settings.encoding_plan
         output_path = output_dir / f"last_clip_processed.{plan.container}"
         ctx = create_assembly_context(self.settings, self.prober, clips, target_w, target_h)
 
-        decoder = _make_decoder(
+        decoder = make_decoder(
             clip,
             len(clips) - 1,
             target_w,
